Refactor the replica node and adding integration test
